Complete project Contains @OneToMany joins

+++ src/main/java/com/qa/business/service/ClassroomService.java
@@ -4,6 +4,6 @@ public interface ClassroomService {
 	String getAll();
 //	String getAClassroomTrainer();
 	String createClassroomTrainer(String trainer);
-	String deleteClassroomTrainer(Long id);
-	String updateClassrooomTrainer(Long id, String trainer);
+	String deleteClassroomTrainer(Integer id);
+	String updateClassrooomTrainer(Integer id, String trainer);
 }
