+++ src/main/java/com/qa/persistence/repository/ClassroomTrainerRepository.java
@@ -5,7 +5,7 @@ public interface ClassroomTrainerRepository {
 	String getAll();
 //	String getAClassroomTrainer();
 	String createClassroomTrainer(String trainer);
-	String deleteClassroomTrainer(Long id);
-	String updateClassrooomTrainer(Long id, String trainer);
+	String deleteClassroomTrainer(Integer id);
+	String updateClassrooomTrainer(Integer id, String trainer);
 
 }
